Fix add form on mobile.  Add remote logging server.
The add to queue form was not correctly closing.  JQM was closing the
dialog then navigating back in the history, leaving the user on the page
he was visiting _before_ coming to the site.  Updating JQM to bleading
edge and setting the dialog to data-role="dialog", fixed the issue.

A scripting issue was causing the add to queue form to not correctly
submit on mobile, now fixed.

Page changes (dialog opening) is still too slow on mobile (iPhone 4).

The remote logger allows for debugging output to be pushed to a remote
shell when necessary (mobile) using $.log.