Add queue time updating and sorting.
The TimeList classes in time_list are now used to insert the section
tags and update the times listed in the queue.  The time should be in
the only time tag in the element, and must have defined data-minutes
(the number of minutes elapsed) and datetime (the starting time in ISO
format) attrs.

To make section sorting more efficient and simpler, when the times are
updated, the section tags are moved, rather than moving the elements
into new sections.  Elements are sorted when added (using the insert
method), so resorting should not be required.