Update isotope queue to sniff browser to decide if css transforms should be used.
Modernizer or the like is not appropriate, as it is already being used
inside isotope to find if transforms are possible, the question is
whether they are advisable.  They prevent us from using relative sizing
which is necessary to allow smooth resizes where the number of columns
don't change, but are required for smooth transitions on the iPad and
other tablets with h/w accel.